Add tests for IP code.
Split IP code into its own file and add tests for isRoutable().
