+++ manager.go
@@ -19,6 +19,8 @@ import (
 	"github.com/miekg/dns"
 )
 
+// Node represents a single node on the Decred network. This struct is json
+// encoded to be stored on disk.
 type Node struct {
 	Services        wire.ServiceFlag
 	LastAttempt     time.Time
@@ -61,60 +63,6 @@ const (
 	// considered dead.
 	pruneExpireTimeout = time.Hour * 8
 )
-
-var (
-	// rfc1918Nets specifies the IPv4 private address blocks as defined by
-	// by RFC1918 (10.0.0.0/8, 172.16.0.0/12, and 192.168.0.0/16).
-	rfc1918Nets = []net.IPNet{
-		ipNet("10.0.0.0", 8, 32),
-		ipNet("172.16.0.0", 12, 32),
-		ipNet("192.168.0.0", 16, 32),
-	}
-
-	// rfc3964Net specifies the IPv6 to IPv4 encapsulation address block as
-	// defined by RFC3964 (2002::/16).
-	rfc3964Net = ipNet("2002::", 16, 128)
-
-	// rfc4380Net specifies the IPv6 teredo tunneling over UDP address block
-	// as defined by RFC4380 (2001::/32).
-	rfc4380Net = ipNet("2001::", 32, 128)
-
-	// rfc4843Net specifies the IPv6 ORCHID address block as defined by
-	// RFC4843 (2001:10::/28).
-	rfc4843Net = ipNet("2001:10::", 28, 128)
-
-	// rfc4862Net specifies the IPv6 stateless address autoconfiguration
-	// address block as defined by RFC4862 (FE80::/64).
-	rfc4862Net = ipNet("FE80::", 64, 128)
-
-	// rfc4193Net specifies the IPv6 unique local address block as defined
-	// by RFC4193 (FC00::/7).
-	rfc4193Net = ipNet("FC00::", 7, 128)
-)
-
-// ipNet returns a net.IPNet struct given the passed IP address string, number
-// of one bits to include at the start of the mask, and the total number of bits
-// for the mask.
-func ipNet(ip string, ones, bits int) net.IPNet {
-	return net.IPNet{IP: net.ParseIP(ip), Mask: net.CIDRMask(ones, bits)}
-}
-
-func isRoutable(addr net.IP) bool {
-	for _, n := range rfc1918Nets {
-		if n.Contains(addr) {
-			return false
-		}
-	}
-	if rfc3964Net.Contains(addr) ||
-		rfc4380Net.Contains(addr) ||
-		rfc4843Net.Contains(addr) ||
-		rfc4862Net.Contains(addr) ||
-		rfc4193Net.Contains(addr) {
-		return false
-	}
-
-	return true
-}
 
 func NewManager(dataDir string, defaultPort string) (*Manager, error) {
 	err := os.MkdirAll(dataDir, 0700)
